Former-commit-id: e30b7b21c7e68f3dfc265dc9518715c98f3ce567 [formerly e30b7b21c7e68f3dfc265dc9518715c98f3ce567 [formerly 34af4107ec9a98bbf386535e63bb4876391aaba3]] [formerly 431c87fd41bb18308d7733d6e8618cbff4917a5c] [formerly e30b7b21c7e68f3dfc265dc9518715c98f3ce567 [formerly e30b7b21c7e68f3dfc265dc9518715c98f3ce567 [formerly 34af4107ec9a98bbf386535e63bb4876391aaba3]] [formerly 431c87fd41bb18308d7733d6e8618cbff4917a5c] [formerly c1ec7ca271ad2d0e2d8057d14c891d50fe9301f6 [formerly c1ec7ca271ad2d0e2d8057d14c891d50fe9301f6 [formerly 6970a8ab466b63cc73f04bce4b53af4e778d76cd]] [formerly 431c87fd41bb18308d7733d6e8618cbff4917a5c [formerly 431c87fd41bb18308d7733d6e8618cbff4917a5c [formerly 8955b95e1587317abef2b1ca2a15ea801acd055a]]]]]
Former-commit-id: c1ec7ca271ad2d0e2d8057d14c891d50fe9301f6 [formerly c1ec7ca271ad2d0e2d8057d14c891d50fe9301f6 [formerly 6970a8ab466b63cc73f04bce4b53af4e778d76cd]]
Former-commit-id: 01ea5ec2b9476702920bb539422942b48c4a424c [formerly b9c44ffaaf9cf640c594f86907e364c03442b553] [formerly 45c4727c18893498f4aa75b050a7de65006a2f63 [formerly 2f15451195d4e1641136355b3eec0afc6ea9db89]] [formerly c1ec7ca271ad2d0e2d8057d14c891d50fe9301f6]
Former-commit-id: a4876c6aa545911681719547fb8db7203baa3951 [formerly 495801399df2fe0e2e29d8bb562963eb003edd97] [formerly c6d429342ae1ce3f2d1d6d9e26e6d71eb21e7845 [formerly 89d956228c8568225b227682c3c5207e520bce78]]
Former-commit-id: 8fdcb1b61bcca4b592de53b59c0eff562d06a9c9 [formerly 6ee1d469b5c1ff85cf2d595a2980c5c0416aa782]
Former-commit-id: 01d4bda7332c7562d62161f888adb39f74483c1e

+++ ncep/gov.noaa.nws.ncep.edex.plugin.ncuair/unit-test/gov/noaa/nws/ncep/edex/plugin/ncuair/util/NcUairParserTest.java
@@ -5,32 +5,16 @@
  *
  * L. Lin       04/09   Creation
  * S. Gurung    09/11  Renamed H5 to Nc and h5 to nc
+ * D. Johnson   08/12  Upgrade to JUnit 4.10.
  * </pre>
  *
  */
 package gov.noaa.nws.ncep.edex.plugin.ncuair.util;
 
-import static org.junit.Assert.*;
-import java.util.Calendar;
-import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+
 import org.junit.Before;
-
-import gov.noaa.nws.ncep.common.dataplugin.ncuair.NcUairLiftedIndex;
-import gov.noaa.nws.ncep.common.dataplugin.ncuair.NcUairMaxWind;
-import gov.noaa.nws.ncep.common.dataplugin.ncuair.NcUairObsLevels;
-import gov.noaa.nws.ncep.common.dataplugin.ncuair.NcUairTropopause;
-import gov.noaa.nws.ncep.common.dataplugin.ncuair.NcUairRecord;
-import gov.noaa.nws.ncep.edex.plugin.ncuair.util.NcUairParser;
-import gov.noaa.nws.ncep.edex.plugin.ncuair.util.NcUairWindGroup;
-import gov.noaa.nws.ncep.common.tools.IDecoderConstantsN;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.*;
-
-import gov.noaa.nws.ncep.edex.tools.decoder.MndTime;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.TimeZone;
+import org.junit.Test;
 
 public class NcUairParserTest {
 	String report = "USUS41 KLWX 190000\r\r\n" + 
@@ -68,7 +52,7 @@ public class NcUairParserTest {
 	@Test
 	public void testGetUairType() {
 		Integer uairType = NcUairParser.getUairType("TTAA");
-		assertEquals( 1, uairType);
+        assertEquals(1, uairType.intValue());
 	}
 }
